app/requirements.txt
psycopg2-binary
pandas
pyyaml
beautifulsoup4
python-dotenv